changed int to double in Point

+++ OOP/Exercises/Solutions/Lecture4.java
@@ -70,7 +70,7 @@ class Point {
     private double x;
     private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
